Expected operation, inconsistent Addr
Cpu works as expected except that I find myself
Using direct addressing to save memory and counting
errors. When I am really confident, it will be simple
to restore indirection to Load, goto, and so on.

However, the compiler is giving me the troubling news
that it thinks the Accumulator is storing a String for
some reason. I never put anything but oh.