Now all indirect addressing, maybe too much
As indicated earlier, I wanted to make addressing consistent
across instructions. A fetch gets the value pointed to in the address
of the instruction. It seemed appropriate for a goto to do the same.
However, in trying to make an ISA program (just the first, ie 12+13)
I found the indirection too clever by half. I think that this
implementation is indirecting twice, which is extremely inconvenient.

Nevertheless, I'll sleep on it, look up the spec's definitions for the
goto/load behavior. Essentially, this commit renders the cpu difficult
to use and has neither program produces its intended effect. I've only
uploaded these because it Might be the indirection in the spec and I'm
just not being careful enough about the isa values.