All addressing direct. Expected behavior.
Reread spec, all addressing should be direct. Cpu rewritten to reflect
that demand. HD and hd2 rewritten and behaved as expected. I'll consider
this version of the Cpu stable until I have finished the simple compiler.
The only likely change that I would make would be to increase the RAM
and word size.