Postfixer compiles, aberrant behavior
Made stack print its contents. I decided to keep the stack class
since Python does not (appearantly) have a peek function for lists.

On the road to the simple compiler, I am implementing Deitel's
PostFix algorithm, which the compiler will use in variable assign-
ments. I committed the sin of recopying the code so one can be
verbose for debugging and the other silently work. (Obviously,
using IDLE would disincentivize this.) Once I am certain it works
as expected, I will delete the verbose mode. CPU gets to keep it
since seeing its state change is one aspect of learning from it.
Frankly, I should make it a bit less verbose for obvious stuff.

As far as the unexpected behavior, I am expecting the answer to the
sample equation to be -10.6 or so but it prints 4. I need to go to
bed, so I've not even worked out on paper how it's reaching that
answer. Probably, I implemented the higher precedence loop for
operators incorrectly (specifically for parenthesis).

Also, Deitel's spec suggests that the temp stack should empty
before infix overflows. However, I was overflowing every time,
so I used its length as a sentinel. I may not be popping enough
during the higher precedence loop? Maybe tomorrow.