Stack fixed, PostEval fixed, ToPost aberrant
It turns out that I mishandled creating the stack class, which is a danger
that I had always expected. But, as peek is so integral to this operation,
the toll had to be paid or I would pay with list deletions.

On the basis of only the sample equation already postfixed, postFixEval
appears to work. I will do more testing later. However, I forgot to test
postFix with the new stack and a last minute run shows it fouled the
conversion, which doesn't surprise me. I am fairly sure that the prece-
dence test is wrong, so I will look at Dijkstra's version to see what I
did wrong. Likely, ( and ) should be equal weight.

Once these correctly interpret arbitrary equations, I will work on the
code generator aspect of the compiler before tackling the symbol map. Yay.