Began compiler, broke ram
I wrote the skeleton for the Simple compiler. I am undecided about how to
handle code generation in terms of whether to make that a separate funct-
ion to do so. I consider it obvious but would need to pass in an uncons-
cionable number of variables. I may need to make some of these global,
like the SML instruction codes, as they are static.

I didn't bother compiling since I know it doesn't at the moment and will
complain about snitty aspects of the python I decided not to comment.

I decided that Ram, not comp should be responsible for loading itself with
SML instructions. However, I am not sure if using except rather than
finally is tripping the try that opens the file. This bears actual study.

Another aspect related to that is a gap in the Deitel spec. It skirts
explaining where the generated code is placed. It refers to an SML array
(Java focus) but never instructs to actually create one, unlike the flag
and symbolTable arrays. So, I am inclined to use Ram itself rather than
just another list. However, doing so gives me the sneaky ability to begin
the simpletron without exporting the instructions to an sml file at all.
Of course, that would make it less useful once I end the program suite as
nothing will be saved.

In that case though, I would need to add an exportToHD function to Ram.
Deitel's omission of that potential suggests that they don't expect me to
use the simpletron's ram for storing the tentative instructions. So, I
guess I shouldn't. But, exportToHD could be useful if I ever try a rudim-
entary OS for it. Almost certainly not, given the whole project's extreme
simplicity. I've seen Warford's Pep8 & Knuth's M/Mix. MMix is probably the
better candidate anyway.