Conditional goto ready, let remains
While it is still a big, big function, the conditional jump is ready for
(eventual) testing. Some repeated aspects were extracted to keep the size
down. Some were not. Search & Reserve symbol now exclusively return the
symbolTable index of the desired symbol (or -1). Locations are extracted
from there.

I also changed the remark command back to rem. In typing the Deitel
examples, I noticed that the full word made comment lines wordy and
the same size as the other commands. Rem is shorter by a character &
stands out a bit. Frankly, I'd prefer to make it something like xx.
But, in this case, I'll defer to the spec.

The change to testCompiler is minor, but reflects intuition that jived
with common practice. Namely, report failure, not success. The Dec 2011
Pragmatic Programmer leaflet had a nice article about testing. That's
pretty much all there is to say. While I should keep building the tests
to match behavior, I'll not break flow with tedium. Deferring tedium
until the end is a dangerous practice but the compiler is the goal, not
practicing unassisted testing.