Debugging continues, deep problems uncovered
Closed the cycle for simple testing even though it can lead to problems.
In this case, it seems almost all the addresses pointed at are incorrect.
In the case of goto, I strongly suspect the problem lies in the second
pass, so I'll check that the lineFlags is handled to spec. The others
are more subtle but just as bad. Describing it is disheartening, but
debugging is the pleasant challenge.
I must bite the  bullet now and fill testCompiler with streams of expec-
ted behaviors.