Core dump on syntax error
Added the printing faculties for smlData, symbolTable, and line flags
whenever a syntax error crops up. Glad I did because the Symbol table just
had 17 copies of the same entry : sy53-line-@ic12. That is particularly
odd because the core dump revealed that 3 & 5 were stored in sml data. Is
that possible without a symbol table entry? Probably. I will need to check
harder whether symbolTable really is this bad or the printing is corrupt.