Cpu output restored
The problem was a) I didn't renew the commented outputs in compiler
(though I don't remember seeing this kind of aberrance at the time.) which
meant I had to test an old version to see that the breaking change was
moving the verbosity to self.opVerbs.

When moving to a dict, I underemphasized caution about when opVerbs
created those strings. Unsurprisingly, it created them with the initial
values and did not update them.

So, it was a simple matter to pull out the statements and shove them into
a separate elif field. While it nicely encapsulates what is going on, I
don't forsee introducing something similar in the simple compiler.

The optimization isn't quite complete. I've been working on it, but I felt
unbricking the cpu demanded turnaround on completion rather than waiting
for an unrelated, but important, feature.