Simple assignment resolved
Would have preferred to finish the optimization, but that's starting to
be a habitual admission. Uploaded because the compiler now accounts for a
let of the form x = 3. Ie, where it doesn't need to use postfix. However,
I have overspecified the postEval(), so I need to fix it. Which means more
play with testPost, which is why it's here again.

Both it and testCompiler now respond to a -v flag when run, to activate
verbose mode. Postfixer is also less obnoxiously verbose now. While I'd
cite its obnoxiousness in leaving it off for the compiler, the reality is
that convert silently and verbosely ARE different. Silence expects a list
and verbose expects a string. I could unite the implementations but,
really seeing it as verbose is more appropriate during development or if
you are learing to use it for the first time. Of course, I have been
fouling assignment() exactly because I don't remember how the shunting
yard and evaluation work.