Improved conditionalProduction()
Before now, I had let the runtime decrement the comparison target for a >=
or <= conditional. Now, constants are decremented at compile time. I also
decrement uninitialized variables, but would need to take runtime and
space to decrement a variable that has been changed.

From here out, I'll try to use emacs primarily. Not because it's awesome
for python, but because it's good for clojure - the next language I'll
practice with. General consensus is that its easier to learn a language or
a tool than both at once. Luckily, I'm returning to emacs after a short
stint preparing for a competition last year.

Cygwin's vim is inferior to mysgwn's vim though. I'll have to look at the
actual commands I guess. Or maybe git's vim on mysgwn is more forgiving
than git's vim on cygwin. Plus, color highlighting.

Next time I'll wrap up the let optimization and push foreward with the for
loop and gosub (which I've been working on offline in the intervening
week).