Still working on optimization
I would upload it but, with uncommented pseudo, I'd be breaking the build.
I'll pretty much finish that tomorrow. I had reversed the application
order, largely by misremembering how reverse polish operations are
applied.

As for testCompiler, it now tells you if you forgot the filename argument
(as I have occasionally).

The for/gosub forward references are going to bloat the complexity of this
monolithic compiler. Conflicted, refactor the structure of mainline or
build interpreter/assembler to handle the complexity from the beginning?
The trick is that those branches will never merge back.