For loop continues, moved old files to testMat
Am converting the pseudocode to real code for the loop. It checks syntax
but nothing else. In doing so, I saw it was dead simple to add an implicit
step of one, as Deitel suggested. Delay because of flagging interest and
some separate work I am uploading to bitbucket instead.

Moved a number of simple files to testMateria. Some are important to have
at front (monkey), but until I'm done, they are clutter. Added ToC because
there are too many functions now for me to remember all the helper
functions. Honestly, I sensed this while fixing the let, and probably
reimplemented the saveVariable when I didn't need to. Despite the tedium
of synchronization, having a list of the function prototypes (that I
indented to reflect invocation heirarcy) was a nice aspect of C++.