v1.0 to canonical specification
That old 1.0 was premature since I broke the cpu. This is the real 1.0.
Any change I make (and some I_just left in) will make it increasingly
uselsess as a reference for an aspiring implementer. Further, this is the
most useful version to describe on the wiki as it meets the spec. From
here, I'll branch to an improved version, with most of their extensions
(loops subroutines strings) but a better design. And another branch for an
interpreter version.