Now with Deitel mode
Deitel's actual spec prompts the student to input the program by hand at
the console. As that's tedious, I skipped immediately to file input. In
looking over the wiki files, I noticed the Simpletron driver was an empty
stub. So, I rounded out the driver with the official input. It's inactive,
of course, but nonetheless present.

All the '.sml' files were renamed to .asm so Github doesn't think they are
Standard ML files.