Naive tests passed, need antagonist tests
Built some tests inside assembler that test the various successful cases
of the assembly syntax. I need to test that the checks for bad syntax
succeed in catching what I want. The trick will be in not letting the bad
syntax halt operation. I don't anticipate much trouble.

After that, I will need to test this with the file operation. It will
reveal if the renaming scheme works. (It should, it is in compiler and I
used it in the py's interpreter.) Then I'll spin up some real examples & a
monkey wrench in assebly for translation. Disassembler should reveal the
intended codes.