Testing done, py2 compliant
Finished the testing I anticipate at this stage. That's good syntax, and
checking if it caught errors thrown by bad syntax. I also verified the
file output. As far as simpletron is concerned, this is it for assembler
presently.

The other evaluator, however, requires it to be py3 compliant. Luckily, I
installed it on Cygwin, so I don't need to bother here. I hope it's just
the print statements, but I'll meet with something unexpected - obviously.
He probably would also appreciate a better view of the program internals
as it runs. He will receive the sml codes as output, but that's a bit
opaque. So, I'll have to either flood this with prints, keep it to select
areas (cleanup_refs()), or print it like in compiler, by sending generic
status function an appropriate string.

Of course, SCompiler will need to emit this intermediate assembly instead
of sml codes. Strictly speaking, I could have it perform either, based on
input, but that's what master is for. I do admit, though, that none of
these faculties is portable. Most other vms have registers & various
addressing modes. I would be working in an underutilizing process if I
retargeted asm to dcpu10 or mips, and that assembler would need revisions
to the compiler to handle the larger grammer interface.