Java Simpletron begun
If I must practice Java, I need to start something from scratch as well as add functionality or refactor other's work. So, as I worked on that verboten html spec, I might as well spe( sinning vim) see how hard it is. And it was pretty hard. Not because the java was unintelligible, but because I couldn't get these to run on bare consoles (cmd or cygwin). At least Eclipse worked after I fixed the rosetta code snippet for buffered reader. (It looped while .ready(), which it never was or always was. whatever) It's late now, so more tomorrow.c