Pseudocode for this direction
As happens, this is a commit to work elsewhere rather than improvement. However, it has a retyping of an aggressively message oriented (for me, anyway) pseudocode to design the java simpletron around.

I also put in a premature flag enum to keep simpletron more of a library (for running later) rather than a static self-starter.