Isa & ram tests
Over the weekend I made the isa enum to deal with the earlier failing of modulus math to handle separating instruction digits from address digits. Much cleaner now.

I also added the smlerror enum. This might be a heavy enum project. This became a concern at line 23 of smlmem. readInt seems like it should return an int or throw an exception if the address is bad. I could trust the cpu, but maybe I shouldn't as that enshrines a bad expectation? Which half of the glass am I showing off?

The tedious half, as I made tedious tests for smlmem. These are obvious how to test, but I am more concerned with cpu.

In the even farther future, I'm concerned with moving from a simulation to a gui simulation based on the MOVE paradigm. Appearantly all these functions separate from the program state. But that's just to scare myself. oh well.